Rollback moving code into graphics_util.h
