+++ engine/graphics/src/graphics_private.h
@@ -19,7 +19,6 @@
 #include <dlib/mutex.h>
 #include <dlib/index_pool.h>
 #include "graphics.h"
-#include "graphics_util.h"
 
 namespace dmGraphics
 {
