+++ engine/graphics/src/graphics_vertices.cpp
@@ -13,7 +13,6 @@
 // specific language governing permissions and limitations under the License.
 
 #include "graphics.h"
-#include "graphics_util.h"
 
 #include <dlib/math.h>
