+++ engine/render/src/render/display_profiles.cpp
@@ -22,7 +22,6 @@
 #include <dlib/profile.h>
 #include <dlib/hashtable.h>
 #include <dlib/utf8.h>
-#include <graphics/graphics_util.h>
 
 #include "display_profiles.h"
 #include "display_profiles_private.h"
